backend/requirements.txt

flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
tensorflow==2.13.0
pillow==9.5.0
numpy==1.23.5
matplotlib==3.5.2
scikit-learn==1.0.2